feat: add multimodal support across online backend
